DROTH-3540 use numeric for m-values in temp_road_address_info table

+++ digiroad2-oracle/src/main/resources/db/migration/V1_30__bring_back_temp_road_address.sql
@@ -7,8 +7,8 @@ CREATE TABLE IF NOT EXISTS temp_road_address_info(
     track_code        bigint NOT NULL,
     start_address_m   bigint NOT NULL,
     end_address_m     bigint NOT NULL,
-    start_m_value     bigint NOT NULL,
-    end_m_value       bigint NOT NULL,
+    start_m_value     NUMERIC(1000, 3) NOT NULL,
+    end_m_value       NUMERIC(1000, 3) NOT NULL,
     side_code         numeric(38),
     created_date      timestamp NOT NULL DEFAULT LOCALTIMESTAMP,
     created_by        varchar(128)
